moving sample filter transformation to shared class
