+++ kroxylicious-sample/src/main/java/io/kroxylicious/sample/SampleProduceRequestFilter.java
@@ -6,20 +6,9 @@
 
 package io.kroxylicious.sample;
 
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
-
 import org.apache.kafka.common.message.ProduceRequestData;
 import org.apache.kafka.common.message.ProduceRequestData.PartitionProduceData;
 import org.apache.kafka.common.message.RequestHeaderData;
-import org.apache.kafka.common.record.CompressionType;
-import org.apache.kafka.common.record.MemoryRecords;
-import org.apache.kafka.common.record.MemoryRecordsBuilder;
-import org.apache.kafka.common.record.MutableRecordBatch;
-import org.apache.kafka.common.record.Record;
-import org.apache.kafka.common.record.RecordBatch;
-import org.apache.kafka.common.record.TimestampType;
-import org.apache.kafka.common.utils.ByteBufferOutputStream;
 
 import io.micrometer.core.instrument.Metrics;
 import io.micrometer.core.instrument.Timer;
@@ -27,6 +16,7 @@ import io.micrometer.core.instrument.Timer;
 import io.kroxylicious.proxy.filter.KrpcFilterContext;
 import io.kroxylicious.proxy.filter.ProduceRequestFilter;
 import io.kroxylicious.sample.config.SampleFilterConfig;
+import io.kroxylicious.sample.util.SampleFilterTransformer;
 
 /**
  * A sample ProduceRequestFilter implementation, intended to demonstrate how custom filters work with
@@ -44,13 +34,11 @@ import io.kroxylicious.sample.config.SampleFilterConfig;
  */
 public class SampleProduceRequestFilter implements ProduceRequestFilter {
 
-    private final String findValue;
-    private final String replaceValue;
+    private final SampleFilterConfig config;
     private final Timer timer;
 
     public SampleProduceRequestFilter(SampleFilterConfig config) {
-        this.findValue = config.getFindValue();
-        this.replaceValue = config.getReplaceValue();
+        this.config = config;
         this.timer = Timer
                 .builder("sample_produce_request_filter_transform")
                 .description("Time taken for the SampleProduceRequestFilter to transform the produce data.")
@@ -68,38 +56,15 @@ public class SampleProduceRequestFilter implements ProduceRequestFilter {
      */
     @Override
     public void onProduceRequest(short apiVersion, RequestHeaderData header, ProduceRequestData request, KrpcFilterContext context) {
-        this.timer.record(() ->
-        // We're timing this to report how long it takes through Micrometer
-        applyTransformation(request, context));
+        this.timer.record(() -> applyTransformation(request, context)); // We're timing this to report how long it takes through Micrometer
         context.forwardRequest(header, request);
     }
 
     private void applyTransformation(ProduceRequestData request, KrpcFilterContext context) {
         request.topicData().forEach(topicData -> {
             for (PartitionProduceData partitionData : topicData.partitionData()) {
-                MemoryRecords records = (MemoryRecords) partitionData.records();
-                ByteBufferOutputStream stream = context.createByteBufferOutputStream(records.sizeInBytes());
-                MemoryRecordsBuilder newRecords = createMemoryRecordsBuilder(stream);
-
-                for (MutableRecordBatch batch : records.batches()) {
-                    for (Record batchRecord : batch) {
-                        newRecords.append(batchRecord.timestamp(), batchRecord.key(), transform(batchRecord.value()));
-                    }
-                }
-
-                partitionData.setRecords(newRecords.build());
+                SampleFilterTransformer.transform(partitionData, context, this.config);
             }
         });
-    }
-
-    private ByteBuffer transform(ByteBuffer in) {
-        return ByteBuffer.wrap(new String(StandardCharsets.UTF_8.decode(in).array()).replaceAll(this.findValue, this.replaceValue).getBytes(StandardCharsets.UTF_8));
-    }
-
-    // Reinventing the wheel a bit here to avoid importing from io.kroxylicious.proxy.internal and to improve readability
-    private static MemoryRecordsBuilder createMemoryRecordsBuilder(ByteBufferOutputStream stream) {
-        return new MemoryRecordsBuilder(stream, RecordBatch.CURRENT_MAGIC_VALUE, CompressionType.NONE, TimestampType.CREATE_TIME, 0, RecordBatch.NO_TIMESTAMP,
-                RecordBatch.NO_PRODUCER_ID, RecordBatch.NO_PRODUCER_EPOCH, RecordBatch.NO_SEQUENCE, false, false, RecordBatch.NO_PARTITION_LEADER_EPOCH,
-                stream.remaining());
     }
 }
